+++ test/wasm_worker/wasm_worker_pthread_api_usage.c
@@ -1,0 +1,7 @@
+#include <pthread.h>
+
+int main() {
+  // This program should fail to link with WASM_WORKERS.
+  pthread_mutex_t my_mutex = PTHREAD_MUTEX_INITIALIZER;
+  pthread_mutex_lock(&my_mutex);
+}